feat: Add buttons and js, ui is functional
Got the game to work! Removed the `game` function because I couldn't
figure out how to automatically play multiple rounds, while making it
wait for the user to click a button.

So instead, just setting it so that playRound only starts when user
clicks one of the play class buttons. And then I have an if statement in
place, so that playRound won't run if either the player or computer
already has 5 points.

Also added a reset button for multiple playthroughs without reloading.

Next steps: break up that huge playRound function. I think I can check
for a roundWinner in a separate function.

Also want to add CSS to make it look nice.