wip: Add some css
Started adding CSS. Still in progress. In particular, working on the
bottom half, with the finalResults.

Originally, I had both emojis and text in the 'play' buttons. But I
learned that by placing <p> tags in the buttons, if you clicked on the
emoji or the text, it would not register that you were clicking on the
button.

So, got rid of the text and kept the emoji to make it work right.