feat: Add more CSS to buttons and round results
Button are now larger, have background color when hovered over, and grow
slightly bigger.

Working on finalResults. Need to figure out how to get the finalResults
message to be in the center below the header in the finalResults box.

Also, need to fix up the look of the reset button.